+++ src/main/webapp/app/shared/layout/header/header.tsx
@@ -35,15 +35,6 @@ export default class Header extends React.Component<IHeaderProps, IHeaderState> 
     this.props.onLocaleChange(event.target.value);
   };
 
-  renderDevRibbon = () =>
-    this.props.isInProduction === false ? (
-      <div className="ribbon dev">
-        <a href="">
-          <Translate contentKey={`global.ribbon.${this.props.ribbonEnv}`} />
-        </a>
-      </div>
-    ) : null;
-
   toggleMenu = () => {
     this.setState({ menuOpen: !this.state.menuOpen });
   };
@@ -55,7 +46,6 @@ export default class Header extends React.Component<IHeaderProps, IHeaderState> 
 
     return (
       <div id="app-header">
-        {this.renderDevRibbon()}
         <LoadingBar className="loading-bar" />
         <Navbar dark expand="sm" fixed="top" className="jh-navbar">
           <NavbarToggler aria-label="Menu" onClick={this.toggleMenu} />
