+++ src/main/webapp/i18n/en/fullEvaluate.json
@@ -16,7 +16,7 @@
             "detail": {
                 "title": "Bản đánh giá"
             },
-            "description": "Mô tả",
+            "description": "Tên",
             "result": "Kết quả",
             "teacher": "Giáo viên"
         }
